requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
faiss-cpu==1.7.3
numpy==1.24.3
python-dotenv==1.0.0
assemblyai==0.0.6
jinja2==3.1.2
requests==2.31.0
python-multipart==0.0.6
